chore: Allow nededless borrows
On the way the lib is constructed, the compiler is unable to figure out
that the impl of `Add` and `Mul` that needs to be used (as is the
`const` one is the one done in the module.

Hence why the `&` are used in many places to perform operations.
This means, we need to exclude the clippy checks.